tests/requirements.txt
# Test dependencies
pytest==8.3.5
pytest-cov==6.0.0
pytest-asyncio==0.25.3
pytest-docker==3.1.1
docker==7.1.0
psycopg2-binary==2.9.11
PyYAML==6.0.3
